pyproject.toml
[project]
name = "remixdatabase"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "twilio>=9.10.4",
]
